+++ Riot/AppDelegate.m
@@ -2801,6 +2801,9 @@ NSString *const AppDelegateDidValidateEmailNotificationClientSecretKey = @"AppDe
                     break;
                 }
             }
+            
+            // Do not warn for unknown devices if cross-signing is enabled
+            mxSession.crypto.warnOnUnknowDevices = !RiotSettings.shared.enableCrossSigning;
         }
         else if (mxSession.state == MXSessionStateClosed)
         {